chore(cli-e2e): delete legacy tui-test suite and rename cli-e2e-web to cli-e2e
The Playwright + wterm suite has full parity, so the legacy tui-test stack is
removed and the new suite takes the canonical name. Also lands the post-review
cleanups: trusted-localhost comment on the wterm host, dead __wterm.getText()
removed from the client, stale CLAUDE.md migration framing collapsed into a
single section, and @nx/playwright/plugin dropped from nx.json (we're the only
Playwright project and targets are fully inline in package.json anyway).

- delete apps/cli-e2e/ (legacy) and rename apps/cli-e2e-web/ -> apps/cli-e2e/
- drop orphan @microsoft/tui-test devDependency + .tui-test eslint ignore
- regenerate package-lock.json from scratch to clear npm's stale workspace entry
- update CI + integration workflows for the new path and e2e:integration target
