+++ .github/workflows/ci.yml
@@ -39,6 +39,6 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: playwright-test-output
-          path: apps/cli-e2e-web/test-output
+          path: apps/cli-e2e/test-output
           if-no-files-found: ignore
           retention-days: 7
