--- apps/cli-e2e-web/src/setup/constants.ts
@@ -1,9 +1,0 @@
-export const TEST_REPO =
-  process.env.TEST_REPO ??
-  'kirby-test-runner/kirby-integration-test-repository';
-
-export function testBranchPrefix(): string {
-  const ts = Date.now();
-  const rand = Math.random().toString(36).slice(2, 8);
-  return `e2e-${ts}-${rand}`;
-}
