--- apps/cli-e2e-web/src/setup/git-repo.ts
@@ -1,49 +1,0 @@
-import { mkdtempSync, rmSync } from 'node:fs';
-import { execSync } from 'node:child_process';
-import { join } from 'node:path';
-import { tmpdir } from 'node:os';
-
-export function createTestRepo(): string {
-  const dir = mkdtempSync(join(tmpdir(), 'kirby-e2e-web-'));
-  execSync('git init', { cwd: dir, stdio: 'pipe' });
-  execSync('git config user.email "test@kirby.dev"', {
-    cwd: dir,
-    stdio: 'pipe',
-  });
-  execSync('git config user.name "Kirby Test"', { cwd: dir, stdio: 'pipe' });
-  execSync('git commit --allow-empty -m "initial"', {
-    cwd: dir,
-    stdio: 'pipe',
-  });
-  return dir;
-}
-
-export function cleanupTestRepo(dir: string): void {
-  try {
-    rmSync(dir, { recursive: true, force: true });
-  } catch {
-    /* best effort */
-  }
-}
-
-/**
- * Register cleanup on process exit for module-scope tempdirs
- * (used by integration tests that clone the real test repo once per file).
- */
-export function registerCleanup(dir: string): void {
-  let cleaned = false;
-  const cleanup = () => {
-    if (cleaned) return;
-    cleaned = true;
-    cleanupTestRepo(dir);
-  };
-  process.on('exit', cleanup);
-  process.on('SIGINT', () => {
-    cleanup();
-    process.kill(process.pid, 'SIGINT');
-  });
-  process.on('SIGTERM', () => {
-    cleanup();
-    process.kill(process.pid, 'SIGTERM');
-  });
-}
