+++ apps/cli-e2e/package.json
@@ -6,22 +6,62 @@
   "nx": {
     "name": "cli-e2e",
     "targets": {
-      "//": "e2e target temporarily disabled — parity held by cli-e2e-web (Playwright+wterm). Tests under src/ are kept on disk for reference during the migration; restore the target below to re-enable.",
-      "//e2e": {
+      "e2e": {
+        "executor": "nx:run-commands",
+        "cache": true,
         "dependsOn": [
           {
             "projects": [
               "cli"
             ],
             "target": "build"
+          },
+          {
+            "projects": [
+              "cli-wterm-host"
+            ],
+            "target": "build"
           }
         ],
-        "executor": "nx:run-commands",
+        "inputs": [
+          "default",
+          "^production",
+          {
+            "externalDependencies": [
+              "@playwright/test"
+            ]
+          }
+        ],
+        "outputs": [
+          "{projectRoot}/test-output/playwright/output",
+          "{projectRoot}/test-output/playwright/report"
+        ],
         "options": {
-          "command": "env -u CI -u CONTINUOUS_INTEGRATION -u GITHUB_ACTIONS npx @microsoft/tui-test",
+          "command": "playwright test --grep-invert @integration",
           "cwd": "apps/cli-e2e"
-        },
-        "cache": false
+        }
+      },
+      "e2e:integration": {
+        "executor": "nx:run-commands",
+        "cache": false,
+        "dependsOn": [
+          {
+            "projects": [
+              "cli"
+            ],
+            "target": "build"
+          },
+          {
+            "projects": [
+              "cli-wterm-host"
+            ],
+            "target": "build"
+          }
+        ],
+        "options": {
+          "command": "playwright test",
+          "cwd": "apps/cli-e2e"
+        }
       }
     }
   }
