+++ apps/cli-e2e/src/fixtures/kirby.ts
