+++ apps/cli-e2e/src/session-sort-selection.test.ts
@@ -1,34 +1,28 @@
-import { test, expect } from '@microsoft/tui-test';
-import type { Terminal } from '@microsoft/tui-test/lib/terminal/term.js';
-import { mkdtempSync, mkdirSync, writeFileSync } from 'node:fs';
 import { execSync } from 'node:child_process';
-import { join, resolve } from 'node:path';
+import { mkdtempSync } from 'node:fs';
 import { tmpdir } from 'node:os';
+import { join } from 'node:path';
+import { test, expect } from './fixtures/kirby.js';
 import { registerCleanup } from './setup/git-repo.js';
 import { TEST_REPO } from './setup/constants.js';
 import { sidebarLocator } from './setup/sidebar.js';
+import type { KirbyTerm } from './fixtures/kirby.js';
 
 const hasGhToken = !!process.env.GH_TOKEN;
 
 // ── Module-scope setup ─────────────────────────────────────────────
-const mainJs = resolve('../cli/dist/main.js');
-
 const cloneDir = mkdtempSync(join(tmpdir(), 'kirby-sort-clone-'));
-const fakeHome = mkdtempSync(join(tmpdir(), 'kirby-sort-home-'));
-const logFile = join(tmpdir(), 'kirby-sort-debug.log');
 registerCleanup(cloneDir);
-registerCleanup(fakeHome);
 
 if (hasGhToken) {
-  // 1. Clone sandbox repo (all branches needed for branch picker)
-  //    Use HTTPS URL directly — gh repo clone may default to SSH which can be blocked.
+  // Use HTTPS URL directly — gh repo clone may default to SSH which can be
+  // blocked. Clone all branches (needed for the branch picker).
   const token = process.env.GH_TOKEN;
   execSync(
     `git clone "https://x-access-token:${token}@github.com/${TEST_REPO}.git" "${cloneDir}"`,
     { stdio: 'pipe' }
   );
 
-  // 2. Configure git identity
   execSync('git config user.email "e2e@kirby.dev"', {
     cwd: cloneDir,
     stdio: 'pipe',
@@ -37,81 +31,54 @@ if (hasGhToken) {
     cwd: cloneDir,
     stdio: 'pipe',
   });
-
-  // 4. Write global config with aiCommand: "cat" to avoid needing claude binary
-  const kirbyDir = join(fakeHome, '.kirby');
-  mkdirSync(kirbyDir, { recursive: true });
-  writeFileSync(
-    join(kirbyDir, 'config.json'),
-    JSON.stringify({ aiCommand: 'cat', keybindPreset: 'vim' }),
-    'utf-8'
-  );
 }
 
 // ── Helper: create a session via the branch picker UI ──────────────
 async function createSessionViaBranchPicker(
-  terminal: Terminal,
+  term: KirbyTerm,
   branchFilter: string,
   waitForTitle: string
 ) {
-  terminal.write('c');
-  await expect(
-    terminal.getByText('Branch Picker', { strict: false })
-  ).toBeVisible();
+  await term.type('c');
+  await expect(term.getByText('Branch Picker').first()).toBeVisible();
 
-  terminal.write(branchFilter);
+  await term.type(branchFilter);
+  await expect(term.getByText(`/ ${branchFilter}`).first()).toBeVisible();
 
-  await expect(
-    terminal.getByText(`/ ${branchFilter}`, { strict: false })
-  ).toBeVisible();
+  await term.press('Enter');
 
-  terminal.write('\r');
-
-  // Wait for the branch picker to close before proceeding. The Enter
-  // triggers an async worktree creation; setCreating(false) fires
-  // synchronously, so the branch picker title disappears first.
-  await expect(
-    terminal.getByText('Branch Picker', { strict: false })
-  ).not.toBeVisible();
+  // Branch picker closes first (setCreating(false) is sync); worktree
+  // creation is async. Wait for the picker to close before asserting
+  // the session row.
+  await expect(term.getByText('Branch Picker').first()).not.toBeVisible();
 
   // Wait for the session row in any icon state (selected or not, running or not).
-  await expect(sidebarLocator(terminal, waitForTitle).any()).toBeVisible({
+  await expect(sidebarLocator(term.page, waitForTitle).any()).toBeVisible({
     timeout: 15_000,
   });
 }
 
-// ── Configure tui-test ─────────────────────────────────────────────
-test.use({
-  rows: 60,
-  columns: 120,
-  program: {
-    file: 'node',
-    args: [mainJs, cloneDir],
-  },
-  env: {
-    ...process.env,
-    HOME: fakeHome,
-    TERM: 'xterm-256color',
-    KIRBY_LOG: logFile,
-  },
-});
+test.describe('@integration Session Sort Selection', () => {
+  test.skip(!hasGhToken, 'Requires GH_TOKEN for real GitHub ops');
 
-// ── Tests ──────────────────────────────────────────────────────────
+  test.use({
+    kirbyRepoPath: cloneDir,
+    kirbyConfig: { aiCommand: 'cat', keybindPreset: 'vim' },
+    rows: 60,
+    cols: 120,
+  });
 
-test.when(
-  hasGhToken,
-  'selects correct session after branch picker creation in sorted sidebar',
-  async ({ terminal }) => {
-    // 1. Wait for Kirby to render
-    await expect(
-      terminal.getByText('Approved by You', { strict: false })
-    ).toBeVisible();
+  test('selects correct session after branch picker creation in sorted sidebar', async ({
+    kirby,
+  }) => {
+    // 1. Wait for PR data to load (reviews section appears)
+    await expect(kirby.term.getByText('Approved by You').first()).toBeVisible();
 
     // 2. Create sessions in an order that DIFFERS from PR-sorted order.
     //    Sorted order (desc PR ID): ai-solver(#39), undo(#38), color(#37)
     //    Creation order:            color(#37), ai-solver(#39), undo(#38)
     //
-    //    This makes the raw list indices differ from sorted indices:
+    //    This makes raw indices differ from sorted indices:
     //      Raw (creation order): [color=0, ai-solver=1, undo=2]
     //      Sorted (desc PR):    [ai-solver=0, undo=1, color=2]
     //
@@ -119,38 +86,37 @@ test.when(
     //    Fixed findSortedIndex('undo') = 1 → sorted[1] = undo (CORRECT)
 
     await createSessionViaBranchPicker(
-      terminal,
+      kirby.term,
       'fixture/add-color',
       'Add color support'
     );
-
     await createSessionViaBranchPicker(
-      terminal,
+      kirby.term,
       'fixture/add-ai-solver',
       'Add AI solver'
     );
 
-    // 3. Wait for PR data to load (PR badges should appear)
-    await expect(terminal.getByText('#39', { strict: false })).toBeVisible({
+    // 3. Wait for PR data on the sessions (#39 badge appears)
+    await expect(kirby.term.getByText('#39').first()).toBeVisible({
       timeout: 30_000,
     });
 
-    // 4. Create the third session: fixture/add-undo-feature (PR #38)
-    //    This is the session that exposes the bug.
+    // 4. Create the third session (fixture/add-undo-feature / PR #38).
+    //    This session exposes the bug.
     await createSessionViaBranchPicker(
-      terminal,
+      kirby.term,
       'fixture/add-undo',
       'Add undo feature'
     );
 
     // 5. The selection indicator should be on the newly created session.
     await expect(
-      sidebarLocator(terminal, 'Add undo feature').selected()
+      sidebarLocator(kirby.term.page, 'Add undo feature').selected()
     ).toBeVisible();
 
     // 6. Confirm selection is NOT on the wrong session (color-support).
-    expect(
-      sidebarLocator(terminal, 'Add color support').selected()
+    await expect(
+      sidebarLocator(kirby.term.page, 'Add color support').selected()
     ).not.toBeVisible();
-  }
-);
+  });
+});
