+++ apps/cli-e2e/src/sidebar-auto-hide.test.ts
@@ -1,99 +1,56 @@
-import { test, expect } from '@microsoft/tui-test';
-import { mkdtempSync, mkdirSync, writeFileSync } from 'node:fs';
-import { join, resolve } from 'node:path';
-import { tmpdir } from 'node:os';
-import { createTestRepo, registerCleanup } from './setup/git-repo.js';
+import { test, expect } from './fixtures/kirby.js';
 
-// Isolated HOME so the test controls the kirby config (aiCommand, keybinds).
-function createIsolatedTestEnv() {
-  const dir = createTestRepo();
-  const home = mkdtempSync(join(tmpdir(), 'kirby-autohide-home-'));
-  registerCleanup(dir);
-  registerCleanup(home);
-
-  const kd = join(home, '.kirby');
-  mkdirSync(kd, { recursive: true });
-  writeFileSync(
-    join(kd, 'config.json'),
-    JSON.stringify({
-      aiCommand: 'echo kirby-session-active && sleep 300',
-      keybindPreset: 'vim',
-    }),
-    'utf-8'
-  );
-
-  return { dir, home };
-}
-
-async function typeText(
-  terminal: { write: (s: string) => void },
-  text: string
-) {
-  for (const ch of text) {
-    terminal.write(ch);
-    await new Promise((r) => setTimeout(r, 80));
-  }
-}
-
-const mainJs = resolve('../cli/dist/main.js');
-const env = createIsolatedTestEnv();
+test.use({
+  kirbyConfig: {
+    aiCommand: 'echo kirby-session-active && sleep 300',
+    keybindPreset: 'vim',
+  },
+});
 
 test.describe('Sidebar auto-hide', () => {
-  test.use({
-    rows: 30,
-    columns: 100,
-    program: { file: 'node', args: [mainJs, env.dir] },
-    env: {
-      ...process.env,
-      HOME: env.home,
-      TERM: 'xterm-256color',
-    },
-  });
-
   test('hides on Tab into a session and reappears on Tab out', async ({
-    terminal,
+    kirby,
   }) => {
     const branchName = 'autohide-e2e';
 
     // 1. Empty state
-    await expect(terminal.getByText('(no sessions)')).toBeVisible();
+    await expect(kirby.term.getByText('(no sessions)')).toBeVisible();
 
     // 2. Create a session via the branch picker
-    terminal.write('c');
-    await expect(terminal.getByText('Branch Picker')).toBeVisible();
-    await typeText(terminal, branchName);
-    await expect(
-      terminal.getByText('(new branch)', { strict: false })
-    ).toBeVisible({ timeout: 5_000 });
-    // Let React re-render so useInput closure captures the updated filter
-    await new Promise((r) => setTimeout(r, 2_000));
-    terminal.write('\r');
-
-    // 3. Session is visible in the sidebar
-    await expect(terminal.getByText('Branch Picker')).not.toBeVisible({
+    await kirby.term.type('c');
+    await expect(kirby.term.getByText('Branch Picker')).toBeVisible();
+    await kirby.term.type(branchName);
+    await expect(kirby.term.getByText(/\(new branch\)/).first()).toBeVisible({
       timeout: 5_000,
     });
-    await expect(terminal.getByText(branchName, { strict: false })).toBeVisible(
-      { timeout: 10_000 }
-    );
+    // Let React re-render so useInput closure captures the updated filter.
+    await kirby.term.page.waitForTimeout(2_000);
+    await kirby.term.press('Enter');
 
-    // 4. Tab → PTY starts, focus moves to terminal, sidebar hides
+    // 3. Session is visible in the sidebar
+    await expect(kirby.term.getByText('Branch Picker')).not.toBeVisible({
+      timeout: 5_000,
+    });
+    await expect(kirby.term.getByText(branchName).first()).toBeVisible({
+      timeout: 10_000,
+    });
+
+    // 4. Tab → PTY starts, focus moves to terminal, sidebar hides.
     //    The session name may still be visible as the main pane title, so
-    //    assert on a sidebar-only element (keybind hint "quit") instead of
-    //    the branch name.
-    terminal.write('\t');
+    //    assert on a sidebar-only element (keybind hint "quit") instead.
+    await kirby.term.press('Tab');
     await expect(
-      terminal.getByText('kirby-session-active', { strict: false })
+      kirby.term.getByText('kirby-session-active').first()
     ).toBeVisible({ timeout: 10_000 });
-    await expect(terminal.getByText('quit', { strict: false })).not.toBeVisible(
-      { timeout: 5_000 }
-    );
+    await expect(kirby.term.getByText('quit').first()).not.toBeVisible({
+      timeout: 5_000,
+    });
 
-    // 5. Ctrl+Space exits the terminal pane → sidebar reappears
+    // 5. Ctrl+Space exits the terminal pane → sidebar reappears.
     //    (Tab is forwarded into the PTY when focused on the agent, so the
-    //    exit key is \x00 — see useRawStdinForward.ts)
-    terminal.write('\x00');
-    await expect(terminal.getByText('quit', { strict: false })).toBeVisible({
+    //    exit key is \x00 — see useRawStdinForward.ts.)
+    await kirby.term.write('\x00');
+    await expect(kirby.term.getByText('quit').first()).toBeVisible({
       timeout: 5_000,
     });
   });
