+++ apps/cli-e2e/tsconfig.json
@@ -1,9 +1,10 @@
 {
   "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "types": ["node"]
-  },
-  "include": ["src/**/*.ts", "tui-test.config.ts"],
-  "exclude": [".tui-test"]
+  "files": [],
+  "include": [],
+  "references": [
+    {
+      "path": "./tsconfig.e2e.json"
+    }
+  ]
 }
