--- apps/cli-e2e/tui-test.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from '@microsoft/tui-test';
-
-export default defineConfig({
-  retries: 0,
-  timeout: 120_000,
-  workers: 1,
-  testMatch: 'src/**/*.test.ts',
-  use: {
-    rows: 30,
-    columns: 100,
-  },
-});
