+++ package.json
@@ -8,7 +8,6 @@
   },
   "private": true,
   "devDependencies": {
-    "@microsoft/tui-test": "^0.0.1",
     "@nx/devkit": "22.5.2",
     "@nx/esbuild": "22.5.2",
     "@nx/eslint": "22.5.2",
